requirements.txt
fastapi==0.111.1
uvicorn==0.24.0
ultralytics==8.1.3
pillow==9.5.0
jinja2